requirements.txt
flask==2.3.3
sendgrid==6.10.0
python-dotenv==1.0.1
